frontend: Split `schemas.ts` into separate files
