+++ frontend/src/common/schemas/index.ts
@@ -1,1 +1,10 @@
-export * from "./schemas";
+export * from "./apartment";
+export * from "./common";
+export * from "./developer";
+export * from "./enums";
+export * from "./functions";
+export * from "./housingCompany";
+export * from "./improvements";
+export * from "./indices";
+export * from "./owner";
+export * from "./propertyManager";
